begin writing invocation; inits buffer
Fixed an issue where commands were not being processed
properly on invocation. Have to, for now, indicate a
command to be executed, as opposed to a file to open, with
an '@' symbol instead of '!', whidh is interpreted by the
shell as a history wildcard and gave the wrong
result.
Working great now! pass it a file or command (preceded by
'@' and quoted for any arguments) and red will print the
file contents or command output, respectively.

TODO: My current goal is to get the print function working;
then add functionality from there.