fix m behavior; 0/default addr; insert/append;
A lot of changes here and there.
Set it so that 0 is not a real address; we use it to carry
default behavior to an operation; each operation can decide
how to use it (each operation has its own defaults)
In some cases, a 0 address may be used, but it is converted to
the best appropriate one (e.g. convert to 1)
The normalize_address function was changed back to doing low-end
normalization... it converts address 0 to 1; we catch empty address
arguments elsewhere and convert to a 0 that is not processed this way
so that it reaches the operation function as a 0 and is known to be
a default address.
Default line adding behavior is append, not insert; this is crucial
due to the fact that we have a 0 address (in a sense) but we do
NOT have a total+1 address. Insert will convert the address to
-= 1 and then append. Address 0 is left as-is, then append.
The move operation was complicated and frustrating, due to the
desire to take it a single line at a time. I managed to get it,
I think, so that it behaves correctly and as desired.