parse invocation; fix 'get_input' bug;
A couple of commits back, I had introduced a change that broke
the program, so that it could not process any commands! I found
the culprit and corrected the issue.

The main addition here is parse_invocation, which does exactly
that and seems to do it well!

Also added some items in todo.

That's it!