whitespace in addrs; change path handling;
This was a bit of a loaded, all over the place update

I started with adding whitespace to the address parsing,
for which I also added several tests. Unfortunately, it wasn't
working! So with all the playing around with the tests, I got
irritated that the test files were still not being cleaned up
but piling up each time and I had to manually delete them. So
this got me going on how paths are handled, and the temp file
and the main saved file, and what happens when you change
the file name or the program does a new buffer save to an
updated file name. That also got me into the file names not
having full paths (how can this be!?) so ultimately I changed
the String values to OsString and Path values, which made for
MUCH cleaner handling of paths and path components, parents,
etc.

After all of this, and all of the changes and fixes I had to
go through to change the way the path and filename variables
(which are EVERYwhere) work and are handled, I finally got
red compiling again and the tests running. Then with quite a
bit more work I got most of the tests passing again - though
still failing on the whitespace addresses.

Finally, getting back to my original intention for this update,
I double checked the regex syntax that the regex crate uses
and... wouldn't you know it, whitespace is "\s" not "\w"!
I made that change and everything just worked.

Seriously? That's all it was? Who the frick uses "\w" for
whitespace anyway? So stupid of me. But hey, I made some
difficult but, I think, worthwhile changes.

Done. Until next time.