implement global operation - amazing!
This is working so well! There are two issues I have found
with this implementation:
* the default print operation that occurs when you leave a
blank line does not work for the last line in the set of
commands - I'm not sure what to do about this but it is pretty
minor
* the current address is altered by the operation, but it
currently doesn't do so according to the OpenBSD manpage... not
quite anyway. Another very small issue that I just didn't feel
like spending more time on; especially since I don't know if
I will ultimately handle that in buf or ops.

I played around with this operation, along with substitutions,
using the global or numbered match substitution, and it just
works perfectly!